requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
google-generativeai
